[UI] BottomNavBar Component 처리
